requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
pydantic==2.12.5
SQLAlchemy==2.0.44
python-dotenv==1.0.1
slowapi==0.1.9
email-validator==2.2.0
psycopg2-binary==2.9.9
